+++ engine/flink/avro-util/src/test/scala/pl/touk/nussknacker/engine/avro/helpers/KafkaAvroSpecMixin.scala
@@ -1,7 +1,6 @@
 package pl.touk.nussknacker.engine.avro.helpers
 
 import com.typesafe.scalalogging.LazyLogging
-import io.confluent.kafka.serializers.KafkaAvroSerializer
 import org.apache.flink.api.common.ExecutionConfig
 import org.apache.flink.streaming.api.scala.StreamExecutionEnvironment
 import org.scalatest.{FunSuite, Matchers}
@@ -16,9 +15,10 @@ import pl.touk.nussknacker.engine.avro.schemaregistry.confluent.client.Confluent
 import pl.touk.nussknacker.engine.avro.schemaregistry.{ExistingSchemaVersion, LatestSchemaVersion, SchemaVersionOption}
 import pl.touk.nussknacker.engine.avro.sink.KafkaAvroSinkFactory
 import pl.touk.nussknacker.engine.avro.source.KafkaAvroSourceFactory
-import pl.touk.nussknacker.engine.build.EspProcessBuilder
+import pl.touk.nussknacker.engine.build.{EspProcessBuilder, GraphBuilder}
 import pl.touk.nussknacker.engine.flink.test.FlinkSpec
 import pl.touk.nussknacker.engine.graph.{EspProcess, expression}
+import pl.touk.nussknacker.engine.kafka.KafkaConfig
 import pl.touk.nussknacker.engine.process.ExecutionConfigPreparer
 import pl.touk.nussknacker.engine.process.ExecutionConfigPreparer.{ProcessSettingsPreparer, UnoptimizedSerializationPreparer}
 import pl.touk.nussknacker.engine.process.registrar.FlinkProcessRegistrar
@@ -29,8 +29,6 @@ import pl.touk.nussknacker.test.{NussknackerAssertions, PatientScalaFutures}
 trait KafkaAvroSpecMixin extends FunSuite with KafkaWithSchemaRegistryOperations with FlinkSpec with SchemaRegistryMixin with Matchers with LazyLogging with NussknackerAssertions with PatientScalaFutures {
 
   import spel.Implicits._
-
-  import collection.JavaConverters._
 
   protected var registrar: FlinkProcessRegistrar = _
 
@@ -54,8 +52,10 @@ trait KafkaAvroSpecMixin extends FunSuite with KafkaWithSchemaRegistryOperations
 
   protected lazy val nodeId: NodeId = NodeId("mock-node-id")
 
-  protected lazy val avroSourceFactory: KafkaAvroSourceFactory[Any] = {
-    new KafkaAvroSourceFactory[Any](schemaRegistryProvider, testProcessObjectDependencies, None)
+  protected def avroSourceFactory(useStringForKey: Boolean): KafkaAvroSourceFactory[Any, Any] = {
+    new KafkaAvroSourceFactory[Any, Any](schemaRegistryProvider, testProcessObjectDependencies, None) {
+      override protected def prepareKafkaConfig: KafkaConfig = super.prepareKafkaConfig.copy(useStringForKey = useStringForKey)
+    }
   }
 
   protected lazy val avroSinkFactory: KafkaAvroSinkFactory = {
@@ -68,6 +68,7 @@ trait KafkaAvroSpecMixin extends FunSuite with KafkaWithSchemaRegistryOperations
     import spel.Implicits._
     val sourceParams = List(TopicParamName -> asSpelExpression(s"'${source.topic}'")) ++ (source match {
       case GenericSourceAvroParam(_, version) => List(SchemaVersionParamName -> asSpelExpression(formatVersionParam(version)))
+      case GenericSourceWithKeySupportAvroParam(_, version) => List(SchemaVersionParamName -> asSpelExpression(formatVersionParam(version)))
       case SpecificSourceAvroParam(_) => List.empty
     })
 
@@ -93,11 +94,15 @@ trait KafkaAvroSpecMixin extends FunSuite with KafkaWithSchemaRegistryOperations
       .map(filter => builder.filter("filter", filter))
       .getOrElse(builder)
 
-    filteredBuilder.emptySink(
-      "end",
-      sink.sinkId,
-      baseSinkParams ++ validationParams ++ sink.valueParams: _*
-    )
+    filteredBuilder
+      .split("split",
+        GraphBuilder.emptySink(
+          "end",
+          sink.sinkId,
+          baseSinkParams ++ validationParams ++ sink.valueParams: _*
+        ),
+        GraphBuilder.sink("outputInputMeta", "#inputMeta", "sinkForInputMeta")
+      )
   }
 
   protected def formatVersionParam(versionOption: SchemaVersionOption): String =
@@ -141,6 +146,10 @@ trait KafkaAvroSpecMixin extends FunSuite with KafkaWithSchemaRegistryOperations
     override def sourceType: String = "kafka-avro-specific"
   }
 
+  case class GenericSourceWithKeySupportAvroParam(topic: String, versionOption: SchemaVersionOption) extends SourceAvroParam {
+    override def sourceType: String = "kafka-avro-key-value"
+  }
+
   object SourceAvroParam {
 
     def forGeneric(topicConfig: TopicConfig, versionOption: SchemaVersionOption): SourceAvroParam =
@@ -148,6 +157,9 @@ trait KafkaAvroSpecMixin extends FunSuite with KafkaWithSchemaRegistryOperations
 
     def forSpecific(topicConfig: TopicConfig): SourceAvroParam =
       SpecificSourceAvroParam(topicConfig.input)
+
+    def forGenericWithKeySchemaSupport(topicConfig: TopicConfig, versionOption: SchemaVersionOption): SourceAvroParam =
+      GenericSourceWithKeySupportAvroParam(topicConfig.input, versionOption)
 
   }