+++ engine/flink/avro-util/src/test/scala/pl/touk/nussknacker/engine/avro/source/KafkaAvroSourceFactorySpec.scala
@@ -6,8 +6,8 @@ import cats.data.Validated.{Invalid, Valid}
 import cats.data.{NonEmptyList, Validated}
 import com.typesafe.config.ConfigFactory
 import io.confluent.kafka.schemaregistry.client.{SchemaRegistryClient => CSchemaRegistryClient}
-import org.apache.avro.Schema
 import org.apache.avro.generic.{GenericData, GenericRecord}
+import org.apache.kafka.clients.consumer.ConsumerRecord
 import pl.touk.nussknacker.engine.api.context.ProcessCompilationError.{CustomNodeError, NodeId}
 import pl.touk.nussknacker.engine.api.context.transformation.{DefinedEagerParameter, OutputVariableNameValue, TypedNodeDependencyValue}
 import pl.touk.nussknacker.engine.api.context.{ProcessCompilationError, ValidationContext}
@@ -17,23 +17,20 @@ import pl.touk.nussknacker.engine.api.{MetaData, StreamMetaData, VariableConstan
 import pl.touk.nussknacker.engine.avro.KafkaAvroBaseTransformer
 import pl.touk.nussknacker.engine.avro.KafkaAvroBaseTransformer.{SchemaVersionParamName, TopicParamName}
 import pl.touk.nussknacker.engine.avro.helpers.KafkaAvroSpecMixin
-import pl.touk.nussknacker.engine.avro.schema.{AvroStringSettings, FullNameV1, FullNameV2}
-import pl.touk.nussknacker.engine.avro.schemaregistry.confluent.ConfluentSchemaRegistryProvider
-import pl.touk.nussknacker.engine.avro.schemaregistry.confluent.client._
-import pl.touk.nussknacker.engine.avro.schemaregistry.confluent.serialization.ConfluentAvroSerializationSchemaFactory
+import pl.touk.nussknacker.engine.avro.schema.{AvroStringSettings, FullNameV1, FullNameV2, PaymentV1}
+import pl.touk.nussknacker.engine.avro.schemaregistry.confluent.client.ConfluentSchemaRegistryClientFactory
 import pl.touk.nussknacker.engine.avro.schemaregistry.{ExistingSchemaVersion, LatestSchemaVersion, SchemaVersionOption}
 import pl.touk.nussknacker.engine.compile.ExpressionCompiler
 import pl.touk.nussknacker.engine.compile.nodecompilation.{GenericNodeTransformationValidator, TransformationResult}
 import pl.touk.nussknacker.engine.flink.api.process.FlinkSourceTestSupport
 import pl.touk.nussknacker.engine.graph.evaluatedparam.Parameter
 import pl.touk.nussknacker.engine.graph.expression.Expression
-import pl.touk.nussknacker.engine.kafka.KafkaConfig
+import pl.touk.nussknacker.engine.kafka.source.InputMeta
 import pl.touk.nussknacker.engine.spel.Implicits._
 import pl.touk.nussknacker.engine.testing.LocalModelData
 import pl.touk.nussknacker.engine.util.process.EmptyProcessConfigCreator
 
 import scala.collection.immutable.ListMap
-import scala.reflect.ClassTag
 
 class KafkaAvroSourceFactorySpec extends KafkaAvroSpecMixin with KafkaAvroSourceSpecMixin {
 
@@ -46,25 +43,25 @@ class KafkaAvroSourceFactorySpec extends KafkaAvroSpecMixin with KafkaAvroSource
   test("should read generated record in v1") {
     val givenObj = FullNameV1.createRecord("Jan", "Kowalski")
 
-    roundTripSingleObject(avroSourceFactory, RecordTopic, ExistingSchemaVersion(1), givenObj, FullNameV1.schema)
+    roundTripValueObject(avroSourceFactory(useStringForKey = true), RecordTopic, ExistingSchemaVersion(1), "", givenObj)
   }
 
   test("should read generated record in v2") {
     val givenObj = FullNameV2.createRecord("Jan", "Maria", "Kowalski")
 
-    roundTripSingleObject(avroSourceFactory, RecordTopic, ExistingSchemaVersion(2), givenObj, FullNameV2.schema)
+    roundTripValueObject(avroSourceFactory(useStringForKey = true), RecordTopic, ExistingSchemaVersion(2), "", givenObj)
   }
 
   test("should read generated record in last version") {
     val givenObj = FullNameV2.createRecord("Jan", "Maria", "Kowalski")
 
-    roundTripSingleObject(avroSourceFactory, RecordTopic, LatestSchemaVersion, givenObj, FullNameV2.schema)
+    roundTripValueObject(avroSourceFactory(useStringForKey = true), RecordTopic, LatestSchemaVersion, "", givenObj)
   }
 
   test("should return validation errors when schema doesn't exist") {
     val givenObj = FullNameV2.createRecord("Jan", "Maria", "Kowalski")
 
-    readLastMessageAndVerify(avroSourceFactory, "fake-topic", ExistingSchemaVersion(1), givenObj, FullNameV2.schema) should matchPattern {
+    readLastMessageAndVerify(avroSourceFactory(useStringForKey = true), "fake-topic", ExistingSchemaVersion(1), "", givenObj) should matchPattern {
       case Invalid(NonEmptyList(CustomNodeError(_, "Fetching schema error for topic: fake-topic, version: ExistingSchemaVersion(1)", _), _)) => ()
     }
   }
@@ -72,33 +69,60 @@ class KafkaAvroSourceFactorySpec extends KafkaAvroSpecMixin with KafkaAvroSource
   test("should return validation errors when schema version doesn't exist") {
     val givenObj = FullNameV2.createRecord("Jan", "Maria", "Kowalski")
 
-    readLastMessageAndVerify(avroSourceFactory, RecordTopic, ExistingSchemaVersion(3), givenObj, FullNameV2.schema) should matchPattern {
+    readLastMessageAndVerify(avroSourceFactory(useStringForKey = true), RecordTopic, ExistingSchemaVersion(3), "", givenObj) should matchPattern {
       case Invalid(NonEmptyList(CustomNodeError(_, "Fetching schema error for topic: testAvroRecordTopic1, version: ExistingSchemaVersion(3)", _), _)) => ()
     }
   }
 
-  test("should read last generated simple object") {
+  test("should read last generated simple object without key schema") {
     val givenObj = 123123
 
-    roundTripSingleObject(avroSourceFactory, IntTopic, ExistingSchemaVersion(1), givenObj, IntSchema)
+    roundTripValueObject(avroSourceFactory(useStringForKey = true), IntTopicNoKey, ExistingSchemaVersion(1), "", givenObj)
+  }
+
+  test("should ignore key schema and empty key value with string-as-key deserialization") {
+    val givenObj = 123123
+
+    roundTripValueObject(avroSourceFactory(useStringForKey = true), IntTopicWithKey, ExistingSchemaVersion(1), "", givenObj)
+  }
+
+  test("should read last generated simple object with expected key schema and valid key") {
+    val givenObj = 123123
+
+    val serializedKey = keySerializer.serialize(IntTopicWithKey, -1)
+    val serializedValue = valueSerializer.serialize(IntTopicWithKey, givenObj)
+    kafkaClient.sendRawMessage(IntTopicWithKey, serializedKey, serializedValue, Some(0))
+
+    readLastMessageAndVerify(avroSourceFactory(useStringForKey = true), IntTopicWithKey, ExistingSchemaVersion(1), new String(serializedKey, StandardCharsets.UTF_8), givenObj)
   }
 
   test("should read object with invalid defaults") {
     val givenObj = new GenericData.Record(InvalidDefaultsSchema)
     givenObj.put("field1", "foo")
 
-    roundTripSingleObject(avroSourceFactory, InvalidDefaultsTopic, ExistingSchemaVersion(1), givenObj, InvalidDefaultsSchema)
+    roundTripValueObject(avroSourceFactory(useStringForKey = true), InvalidDefaultsTopic, ExistingSchemaVersion(1), "", givenObj)
   }
 
-  test("should read last generated key-value object") {
-    val sourceFactory = createKeyValueAvroSourceFactory[Int, Int]
-    val givenObj = (123, 345)
+  test("should read last generated key-value object, simple type") {
+    val givenKey = 123
+    val givenValue = 456
 
-    val serializedKey = keySerializer.serialize(IntTopic, givenObj._1)
-    val serializedValue = valueSerializer.serialize(IntTopic, givenObj._2)
-    kafkaClient.sendRawMessage(IntTopic, serializedKey, serializedValue, Some(0))
+    val serializedKey = keySerializer.serialize(IntTopicWithKey, givenKey)
+    val serializedValue = valueSerializer.serialize(IntTopicWithKey, givenValue)
+    kafkaClient.sendRawMessage(IntTopicWithKey, serializedKey, serializedValue, Some(0))
 
-    readLastMessageAndVerify(sourceFactory, IntTopic, ExistingSchemaVersion(1), givenObj, IntSchema)
+    roundTripKeyValueObject(avroSourceFactory(useStringForKey = false), IntTopicWithKey, ExistingSchemaVersion(1), givenKey, givenValue)
+  }
+
+  test("should read last generated key-value object, complex object") {
+    val givenKey = FullNameV1.record
+    val givenValue = PaymentV1.record
+
+    val serializedKey = keySerializer.serialize(RecordTopicWithKey, givenKey)
+    val serializedValue = valueSerializer.serialize(RecordTopicWithKey, givenValue)
+    kafkaClient.sendRawMessage(RecordTopicWithKey, serializedKey, serializedValue, Some(0))
+
+    roundTripKeyValueObject(avroSourceFactory(useStringForKey = false), RecordTopicWithKey, ExistingSchemaVersion(1), givenKey, givenValue)
   }
 
   test("Should validate specific version") {
@@ -121,7 +145,7 @@ class KafkaAvroSourceFactorySpec extends KafkaAvroSpecMixin with KafkaAvroSource
     result.errors shouldBe
       CustomNodeError("id", "Schema subject doesn't exist.", Some(TopicParamName)) ::
       CustomNodeError("id", "Fetching schema error for topic: terefere, version: LatestSchemaVersion", Some(SchemaVersionParamName)) :: Nil
-    result.outputContext shouldBe ValidationContext(Map(VariableConstants.InputVariableName -> Unknown))
+    result.outputContext shouldBe ValidationContext(Map(VariableConstants.InputVariableName -> Unknown, VariableConstants.InputMetaVariableName -> InputMeta.withType(Unknown)))
   }
 
   test("Should return sane error on invalid version") {
@@ -129,7 +153,7 @@ class KafkaAvroSourceFactorySpec extends KafkaAvroSpecMixin with KafkaAvroSource
       SchemaVersionParamName -> "'12345'")
 
     result.errors shouldBe CustomNodeError("id", "Fetching schema error for topic: testAvroRecordTopic1, version: ExistingSchemaVersion(12345)", Some(SchemaVersionParamName)) :: Nil
-    result.outputContext shouldBe ValidationContext(Map(VariableConstants.InputVariableName -> Unknown))
+    result.outputContext shouldBe ValidationContext(Map(VariableConstants.InputVariableName -> Unknown, VariableConstants.InputMetaVariableName -> InputMeta.withType(Unknown)))
   }
 
   test("Should properly detect input type") {
@@ -137,13 +161,19 @@ class KafkaAvroSourceFactorySpec extends KafkaAvroSpecMixin with KafkaAvroSource
       SchemaVersionParamName -> s"'${SchemaVersionOption.LatestOptionName}'")
 
     result.errors shouldBe Nil
-    result.outputContext shouldBe ValidationContext(Map(VariableConstants.InputVariableName -> TypedObjectTypingResult(
-      ListMap(
-        "first" -> AvroStringSettings.stringTypingResult,
-        "middle" -> AvroStringSettings.stringTypingResult,
-        "last" -> AvroStringSettings.stringTypingResult
-      ), Typed.typedClass[GenericRecord]
-    )))
+    result.outputContext shouldBe ValidationContext(
+      Map(
+        VariableConstants.InputVariableName -> TypedObjectTypingResult(
+          ListMap(
+            "first" -> AvroStringSettings.stringTypingResult,
+            "middle" -> AvroStringSettings.stringTypingResult,
+            "last" -> AvroStringSettings.stringTypingResult
+          ),
+          Typed.typedClass[GenericRecord]
+        ),
+        VariableConstants.InputMetaVariableName -> InputMeta.withType(Typed[String])
+      )
+    )
   }
 
   private def validate(params: (String, Expression)*): TransformationResult = {
@@ -156,38 +186,32 @@ class KafkaAvroSourceFactorySpec extends KafkaAvroSpecMixin with KafkaAvroSource
     implicit val meta: MetaData = MetaData("processId", StreamMetaData())
     implicit val nodeId: NodeId = NodeId("id")
     val paramsList = params.toList.map(p => Parameter(p._1, p._2))
-    validator.validateNode(avroSourceFactory, paramsList, Nil, Some(VariableConstants.InputVariableName))(ValidationContext()).toOption.get
+    validator.validateNode(avroSourceFactory(useStringForKey = true), paramsList, Nil, Some(VariableConstants.InputVariableName))(ValidationContext()).toOption.get
   }
 
-  private def createKeyValueAvroSourceFactory[K: ClassTag, V: ClassTag]: KafkaAvroSourceFactory[Any] = {
-    val deserializerFactory = new TupleAvroKeyValueKafkaAvroDeserializerSchemaFactory[K, V](factory)
-    val provider = new ConfluentSchemaRegistryProvider(
-      factory,
-      new ConfluentAvroSerializationSchemaFactory(factory),
-      deserializerFactory,
-      formatKey = true)
-    new KafkaAvroSourceFactory[Any](provider, testProcessObjectDependencies, None) {
-      override protected def prepareKafkaConfig: KafkaConfig = super.prepareKafkaConfig.copy(useStringForKey = false)
-    }
+  private def roundTripValueObject(sourceFactory: KafkaAvroSourceFactory[Any, Any], topic: String, versionOption: SchemaVersionOption, givenKey: Any, givenValue: Any) = {
+    pushMessage(givenValue, topic)
+    readLastMessageAndVerify(sourceFactory, topic, versionOption, givenKey, givenValue)
   }
 
-  private def roundTripSingleObject(sourceFactory: KafkaAvroSourceFactory[Any], topic: String, versionOption: SchemaVersionOption, givenObj: Any, expectedSchema: Schema) = {
-    pushMessage(givenObj, topic)
-    readLastMessageAndVerify(sourceFactory, topic, versionOption, givenObj, expectedSchema)
+  private def roundTripKeyValueObject(sourceFactory: KafkaAvroSourceFactory[Any, Any], topic: String, versionOption: SchemaVersionOption, givenKey: Any, givenValue: Any) = {
+    pushMessageWithKey(givenKey, givenValue, topic)
+    readLastMessageAndVerify(sourceFactory, topic, versionOption, givenKey, givenValue)
   }
 
-  private def readLastMessageAndVerify(sourceFactory: KafkaAvroSourceFactory[Any], topic: String, versionOption: SchemaVersionOption, givenObj: Any, expectedSchema: Schema) = {
-    createValidatedSource(sourceFactory, topic, versionOption, expectedSchema)
+  private def readLastMessageAndVerify(sourceFactory: KafkaAvroSourceFactory[Any, Any], topic: String, versionOption: SchemaVersionOption, givenKey: Any, givenValue: Any) = {
+    createValidatedSource(sourceFactory, topic, versionOption)
       .map(source => {
         val bytes = source.generateTestData(1)
         info("test object: " + new String(bytes, StandardCharsets.UTF_8))
-        val deserializedObj = source.testDataParser.parseTestData(bytes)
+        val deserializedObj = source.testDataParser.parseTestData(bytes).head.asInstanceOf[ConsumerRecord[Any, Any]]
 
-        deserializedObj shouldEqual List(givenObj)
+        deserializedObj.key() shouldEqual givenKey
+        deserializedObj.value() shouldEqual givenValue
       })
   }
 
-  private def createValidatedSource(sourceFactory: KafkaAvroSourceFactory[Any], topic: String, versionOption: SchemaVersionOption, expectedSchema: Schema) = {
+  private def createValidatedSource(sourceFactory: KafkaAvroSourceFactory[Any, Any], topic: String, versionOption: SchemaVersionOption) = {
     val version = versionOption match {
       case LatestSchemaVersion => SchemaVersionOption.LatestOptionName
       case ExistingSchemaVersion(v) => v.toString
@@ -207,7 +231,7 @@ class KafkaAvroSourceFactorySpec extends KafkaAvroSpecMixin with KafkaAvroSource
   // This transformation can return
   // - the state that contains information on runtime key-value schemas, which is required in createSource.
   // - validation errors
-  private def validateParamsAndInitializeState(sourceFactory: KafkaAvroSourceFactory[Any], topic: String, version: String): Validated[NonEmptyList[ProcessCompilationError], Option[KafkaAvroSourceFactory.KafkaAvroSourceFactoryState]] = {
+  private def validateParamsAndInitializeState(sourceFactory: KafkaAvroSourceFactory[Any, Any], topic: String, version: String): Validated[NonEmptyList[ProcessCompilationError], Option[KafkaAvroSourceFactory.KafkaAvroSourceFactoryState]] = {
     implicit val nodeId: NodeId = NodeId("dummy")
     val parameters = (TopicParamName, DefinedEagerParameter(topic, null)) :: (SchemaVersionParamName, DefinedEagerParameter(version, null)) :: Nil
     val definition = sourceFactory.contextTransformation(ValidationContext(), List(OutputVariableNameValue("dummy")))