+++ engine/flink/avro-util/src/test/scala/pl/touk/nussknacker/engine/avro/source/KafkaAvroSourceSpecMixin.scala
@@ -7,7 +7,7 @@ import pl.touk.nussknacker.engine.api.LazyParameter
 import pl.touk.nussknacker.engine.api.typed.typing
 import pl.touk.nussknacker.engine.avro.{AvroUtils, TestSchemaRegistryClientFactory}
 import pl.touk.nussknacker.engine.avro.encode.{BestEffortAvroEncoder, ValidationMode}
-import pl.touk.nussknacker.engine.avro.schema.{FullNameV1, FullNameV2}
+import pl.touk.nussknacker.engine.avro.schema.{FullNameV1, FullNameV2, PaymentV1}
 import pl.touk.nussknacker.engine.avro.schemaregistry.confluent.client.{CachedConfluentSchemaRegistryClientFactory, MockConfluentSchemaRegistryClientBuilder}
 import pl.touk.nussknacker.engine.avro.typed.AvroSchemaTypeDefinitionExtractor
 
@@ -25,7 +25,9 @@ trait KafkaAvroSourceSpecMixin {
   object KafkaAvroSourceMockSchemaRegistry {
 
     val RecordTopic: String = "testAvroRecordTopic1"
-    val IntTopic: String = "testAvroIntTopic1"
+    val RecordTopicWithKey: String = "testAvroRecordTopic1WithKey"
+    val IntTopicWithKey: String = "testAvroIntTopic1WithKey"
+    val IntTopicNoKey: String = "testAvroIntTopic1NoKey"
     val InvalidDefaultsTopic: String = "testAvroInvalidDefaultsTopic1"
 
     val IntSchema: Schema = AvroUtils.parseSchema(
@@ -54,8 +56,11 @@ trait KafkaAvroSourceSpecMixin {
     val schemaRegistryMockClient: CSchemaRegistryClient = new MockConfluentSchemaRegistryClientBuilder()
       .register(RecordTopic, FullNameV1.schema, 1, isKey = false)
       .register(RecordTopic, FullNameV2.schema, 2, isKey = false)
-      .register(IntTopic, IntSchema, 1, isKey = false)
-      .register(IntTopic, IntSchema, 1, isKey = true)
+      .register(RecordTopicWithKey, PaymentV1.schema, 1, isKey = false)
+      .register(RecordTopicWithKey, FullNameV1.schema, 1, isKey = true)
+      .register(IntTopicNoKey, IntSchema, 1, isKey = false)
+      .register(IntTopicWithKey, IntSchema, 1, isKey = false)
+      .register(IntTopicWithKey, IntSchema, 1, isKey = true)
       .register(InvalidDefaultsTopic, InvalidDefaultsSchema, 1, isKey = false)
       .build